queue_consumer/requirements.txt
pika==1.3.2
psycopg2-binary==2.9.9
SQLAlchemy==2.0.30